Add priority area + show todos for each project
Created a div to be placed before the todo name
to indicate the priority level. Created the
functionality of clicking on a project sidebar
would give you the todos that were specifically
made in that section (ie gave each todo a
property of the section that the user made the
todo in).

Now, I'll work on using localStorage to keep the
information throughout usages and will then
hopefully deal with the date