Implment local storage and fixed circular import
The way that my js files were importing from each
other cuased a circular import that caused an
error when running. To solve this, data.js was
created to hold the main todo list, the main
projects list, and the current project.

Now, I'll work on the Today, Week, and Month tab
to display todos for the respective time frames
using the date-fns npm packages.